pyproject.toml
[project]
name = "employee-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "faker>=37.3.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
]
